+++ rebirth/base/KeepJava/src/com/noob/algorithm/common150/q071/Solution1.java
@@ -1,6 +1,5 @@
 package com.noob.algorithm.common150.q071;
 
-import com.sun.deploy.util.StringUtils;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
