+++ rebirth/base/KeepJava/src/com/noob/algorithm/common150/q082/Solution1.java
@@ -1,0 +1,14 @@
+package com.noob.algorithm.common150.q082;
+
+import com.noob.algorithm.hot100.base.ListNode;
+
+/**
+ * 082 删除链表中的重复元素
+ */
+public class Solution1 {
+
+    public ListNode deleteDuplicates(ListNode head) {
+
+    }
+
+}
