+++ rebirth/base/KeepJava/src/com/noob/algorithm/hot100/q002/Solution.java
@@ -1,6 +1,6 @@
 package com.noob.algorithm.hot100.q002;
 
-import com.noob.algorithm.hot100.q160.ListNode;
+import com.noob.algorithm.hot100.base.ListNode;
 
 /**
  * 两数相加
